requirements.txt
pynput==1.6.8
Pillow==6.2.2
matplotlib==3.0.3
seaborn==0.9.1
numpy==1.16.6
pandas==0.24.2
pystray==0.17.3
pymsgbox==1.0.9
plyer==1.4.3
six==1.14.0
packaging==20.9
pyparsing==2.4.7 